+++ app/src/androidTest/java/nextstep/payments/ui/component/PaymentCardTest.kt
@@ -8,9 +8,9 @@ import nextstep.payments.BaseComposeTest
 import nextstep.payments.R
 import nextstep.payments.ui.component.card.PaymentCard
 import nextstep.payments.ui.screen.newcard.model.BankTypeModel
+import nextstep.payments.ui.utils.chunkedCardNumber
 import nextstep.payments.ui.utils.toFormattedExpirationDate
 import nextstep.payments.utils.assertBackgroundColor
-import nextstep.payments.utils.chunkedCardNumber
 import org.junit.Test
 
 class PaymentCardTest : BaseComposeTest() {
