--- app/src/androidTest/java/nextstep/payments/utils/CardUtils.kt
@@ -1,3 +1,0 @@
-package nextstep.payments.utils
-
-fun String.chunkedCardNumber() = chunked(4).joinToString(" ")
